requirements.txt
## requirements.txt
```
streamlit
openai
pdfplumber
PyPDF2
python-docx
requests
beautifulsoup4
```